Abstracted Material Class and created an Orthographic Camera

+++ src/VertexArray.cpp
@@ -16,7 +16,7 @@ void VertexArray::AddBuffer(const VertexBuffer& vb, const VertexBufferLayout& la
     for(unsigned int i = 0; i < elements.size(); i++){
         const auto& element = elements[i];
         GLCall(glEnableVertexAttribArray(i));
-        GLCall(glVertexAttribPointer(i, element.count, element.type, element.normalized, layout.GetStride(), (const void*)offset));
+        GLCall(glVertexAttribPointer(i, element.count, element.type, element.normalized, layout.GetStride(), (const void*)(uintptr_t)offset));
         offset += element.count * VertexBufferElement::GetSizeOfType(element.type);
     }
 }
